+++ test/dummy/app/controllers/shopify_plugin_demo/uninstalls_controller.rb
@@ -4,6 +4,11 @@ class ShopifyPluginDemo::UninstallsController < ActionController::Base
   skip_forgery_protection
 
   def create
+    unless webhook_signed?
+      head :unauthorized
+      return
+    end
+
     shop_domain = RecordingStudioShopifyPluginTemplate::ShopifySessionClaims.normalize_shop(
       params[:shop] || params.dig(:webhook, :shop_domain) || params.dig(:webhook, :myshopify_domain)
     )
@@ -15,5 +20,22 @@ class ShopifyPluginDemo::UninstallsController < ActionController::Base
       )
     end
     head :ok
+  end
+
+  private
+
+  def webhook_signed?
+    RecordingStudioShopifyPluginTemplate::ShopifyWebhookHmac.valid?(
+      raw_body: request.raw_post,
+      hmac_header: request.headers[RecordingStudioShopifyPluginTemplate::ShopifyWebhookHmac::HEADER],
+      secret: partner_api_secret
+    )
+  end
+
+  def partner_api_secret
+    client_id = RecordingStudioShopifyPluginTemplate.configuration.registered_app_client_id
+    return if client_id.blank?
+
+    RecordingStudioOauth::OauthClient.find_by(client_id: client_id)&.session_token_secret
   end
 end